# docs/openapi-v2.yaml
openapi: 3.1.0
info:
  title: DST Admin API
  version: 2.0.0-alpha.1
  description: >-
    Contract for the authenticated DST Admin management API. JSON fields use
    camelCase. Authenticated state-changing requests require X-CSRF-Token and
    Idempotency-Key headers. Clients send X-DST-Runtime-Target with local or an
    agent-prefixed target ID; unsupported remote domain operations fail closed.
servers:
  - url: /api/v2
...
paths:
...
  /rooms/{roomId}:
    parameters:
      - $ref: "#/components/parameters/RoomID"
    get:
      tags: [Rooms]
      operationId: getRoom
      summary: Read a discovered room
      responses:
        "200":
          description: Room details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Rooms]
      operationId: deleteRoom
      summary: Move a stopped room into the save-root recovery directory
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteRoomInput"
      responses:
        "200":
          description: Room moved to the recovery directory
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteRoomResultEnvelope"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
...
  /rooms/{roomId}/runtime:
    parameters:
      - $ref: "#/components/parameters/RoomID"
    get:
      tags: [Runtime]
      operationId: getRoomRuntime
      summary: Inspect managed customcommands.lua installation and health for every shard
      responses:
        "200": { description: Per-shard runtime status, content: { application/json: { schema: { $ref: "#/components/schemas/DSTRuntimeReportListEnvelope" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
  /rooms/{roomId}/runtime/actions/install:
    parameters:
      - $ref: "#/components/parameters/RoomID"
    post:
      tags: [Runtime]
      operationId: installRoomRuntime
      summary: Idempotently install or upgrade the managed runtime in every shard
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }, { $ref: "#/components/parameters/IdempotencyKey" }]
      responses:
        "200": { description: Runtime installation results, content: { application/json: { schema: { $ref: "#/components/schemas/DSTRuntimeStatusListEnvelope" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
  /rooms/{roomId}/worlds/{worldId}/runtime/actions/install:
    parameters:
      - $ref: "#/components/parameters/RoomID"
      - $ref: "#/components/parameters/WorldID"
    post:
      tags: [Runtime]
      operationId: installWorldRuntime
      summary: Idempotently install or upgrade the managed runtime in one shard
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }, { $ref: "#/components/parameters/IdempotencyKey" }]
      responses:
        "200": { description: Runtime installation result, content: { application/json: { schema: { $ref: "#/components/schemas/DSTRuntimeStatusEnvelope" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
  /rooms/{roomId}/worlds/{worldId}/runtime/backups:
    parameters:
      - $ref: "#/components/parameters/RoomID"
      - $ref: "#/components/parameters/WorldID"
    get:
      tags: [Runtime]
      operationId: listWorldRuntimeBackups
      summary: List verified managed-runtime state backups for one shard
      responses:
        "200": { description: Runtime backups, content: { application/json: { schema: { $ref: "#/components/schemas/DSTRuntimeBackupListEnvelope" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
  /rooms/{roomId}/worlds/{worldId}/runtime/events:
    parameters:
      - $ref: "#/components/parameters/RoomID"
      - $ref: "#/components/parameters/WorldID"
    get:
      tags: [Rooms]
      operationId: getWorldRuntimeEvents
      summary: Read the bounded recent customcommands runtime event batches
      responses:
        "200": { description: Recent runtime events, content: { application/json: { schema: { $ref: "#/components/schemas/DSTRuntimeEventBatchEnvelope" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /rooms/{roomId}/worlds/{worldId}/runtime/diagnostics/latest:
    parameters:
      - $ref: "#/components/parameters/RoomID"
      - $ref: "#/components/parameters/WorldID"
    get:
      tags: [Rooms]
      operationId: getLatestWorldRuntimeDiagnostic
      summary: Read the latest bounded customcommands runtime diagnostic report
      responses:
        "200": { description: Latest runtime diagnostic report, content: { application/json: { schema: { $ref: "#/components/schemas/DSTRuntimeDiagnosticEnvelope" } } } }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /rooms/{roomId}/worlds/{worldId}/runtime/diagnostics:
    parameters:
      - $ref: "#/components/parameters/RoomID"
      - $ref: "#/components/parameters/WorldID"
    post:
      tags: [Rooms]
      operationId: captureWorldRuntimeDiagnostic
      summary: Run one bounded runtime diagnostic and wait for its structured report
      parameters:
        - $ref: "#/components/parameters/CSRFToken"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DSTRuntimeDiagnosticInput" }
      responses:
        "200": { description: Completed runtime diagnostic report, content: { application/json: { schema: { $ref: "#/components/schemas/DSTRuntimeDiagnosticEnvelope" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /rooms/{roomId}/worlds/{worldId}/runtime/actions/rollback:
    parameters:
      - $ref: "#/components/parameters/RoomID"
      - $ref: "#/components/parameters/WorldID"
    post:
      tags: [Runtime]
      operationId: rollbackWorldRuntime
      summary: Restore a verified runtime backup while the shard is stopped
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }, { $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/DSTRuntimeRollbackInput" } } } }
      responses:
        "200": { description: Runtime rollback result, content: { application/json: { schema: { $ref: "#/components/schemas/DSTRuntimeStatusEnvelope" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
  /rooms/{roomId}/worlds/{worldId}/runtime:
    parameters:
      - $ref: "#/components/parameters/RoomID"
      - $ref: "#/components/parameters/WorldID"
    delete:
      tags: [Runtime]
      operationId: uninstallWorldRuntime
      summary: Remove only DST Admin managed runtime content while the shard is stopped
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }, { $ref: "#/components/parameters/IdempotencyKey" }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/DSTRuntimeConfirmationInput" } } } }
      responses:
        "200": { description: Runtime uninstall result, content: { application/json: { schema: { $ref: "#/components/schemas/DSTRuntimeStatusEnvelope" } } } }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/UnprocessableEntity" }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: dst_admin_session
  parameters:
    CSRFToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: A unique operation key. Reusing it with the same request replays the original response; reusing it with different JSON content returns 409.
      schema:
        type: string
        minLength: 8
        maxLength: 128
        pattern: '^[A-Za-z0-9][A-Za-z0-9._:-]{7,127}$'
    RoomID:
      name: roomId
      in: path
      required: true
      schema:
        type: string
    WorldID:
      name: worldId
      in: path
      required: true
      schema:
        type: string
...
  responses:
...
    Conflict:
      description: The request conflicts with current setup state
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    UnprocessableEntity:
      description: The request is valid JSON but violates a domain constraint
      content:
        application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    NotFound:
      description: The requested resource does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
    Meta:
      type: object
      required: [requestId]
      properties:
        requestId:
          type: string
    APIError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details: {}
    ErrorEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data:
          type: "null"
        error:
          $ref: "#/components/schemas/APIError"
        meta:
          $ref: "#/components/schemas/Meta"
...
    Room:
      type: object
      additionalProperties: false
      required:
        - id
        - directoryName
        - name
        - description
        - gameMode
        - maxPlayers
        - pvp
        - passwordProtected
        - managed
        - worldCount
        - updatedAt
      properties:
        id:
          type: string
          description: Canonical opaque ID derived from the safe directory component
        directoryName:
          type: string
        name:
          type: string
        description:
          type: string
        gameMode:
          type: string
        maxPlayers:
          type: integer
        pvp:
          type: boolean
        passwordProtected:
          type: boolean
        managed:
          type: boolean
        worldCount:
          type: integer
        updatedAt:
          type: string
          format: date-time
...
    RoomEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data:
          $ref: "#/components/schemas/Room"
        error:
          type: "null"
        meta:
          $ref: "#/components/schemas/Meta"
...
    DeleteRoomInput:
      type: object
      additionalProperties: false
      required: [confirmation]
      properties:
        confirmation:
          type: string
          description: Must exactly equal the room name
    DeleteRoomResult:
      type: object
      additionalProperties: false
      required: [room, recoveryName]
      properties:
        room:
          $ref: "#/components/schemas/Room"
        recoveryName:
          type: string
          description: Relative path below the configured save root
    DeleteRoomResultEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data:
          $ref: "#/components/schemas/DeleteRoomResult"
        error:
          type: "null"
        meta:
          $ref: "#/components/schemas/Meta"
...
    DSTRuntimeInstallState:
      type: string
      enum: [missing, installed, outdated, invalid]
    DSTRuntimeHealthState:
      type: string
      enum: [ready, starting, stopped, degraded, unavailable]
    DSTRuntimeStatus:
      type: object
      additionalProperties: false
      required: [roomId, worldId, worldName, state]
      properties:
        roomId: { type: string }
        worldId: { type: string }
        worldName: { type: string }
        state: { $ref: "#/components/schemas/DSTRuntimeInstallState" }
        version: { type: string }
        protocolVersion: { type: integer, minimum: 1 }
        changed: { type: boolean }
        backupPath: { type: string, description: Relative path below the shard directory }
        message: { type: string }
    DSTRuntimeHealth:
      type: object
      additionalProperties: false
      required: [schemaVersion, producerVersion, producerInstanceId, sessionId, shardId, running, ready, writing, sequence, consecutiveFailures]
      properties:
        schemaVersion: { type: integer, minimum: 1 }
        producerVersion: { type: string }
        producerInstanceId: { type: string }
        sessionId: { type: string }
        shardId: { type: string }
        running: { type: boolean }
        ready: { type: boolean }
        writing: { type: boolean }
        sequence: { type: integer, format: int64, minimum: 0 }
        lastCapturedAtUnix: { type: integer, format: int64 }
        lastWrittenAtUnix: { type: integer, format: int64 }
        lastDurationMilliseconds: { type: number, minimum: 0 }
        lastError: { type: string }
        consecutiveFailures: { type: integer, minimum: 0 }
        modules:
          type: object
          additionalProperties: { $ref: "#/components/schemas/DSTRuntimeModuleHealth" }
    DSTRuntimeModuleHealth:
      type: object
      additionalProperties: false
      required: [running, ready]
      properties:
        running: { type: boolean }
        ready: { type: boolean }
        busy: { type: boolean }
        sequence: { type: integer, format: int64, minimum: 0 }
        lastError: { type: string, maxLength: 1024 }
    DSTRuntimeEvent:
      type: object
      additionalProperties: false
      required: [sequence, kind, occurredAtUnix, fields]
      properties:
        sequence: { type: integer, format: int64, minimum: 1 }
        kind:
          type: string
          enum: [player.joined, player.left, world.save, world.reset, world.shutdown, world.cycles, world.phase, world.season, world.rain]
        occurredAtUnix: { type: integer, format: int64, minimum: 1 }
        fields: { type: object, additionalProperties: true }
    DSTRuntimeEventBatch:
      type: object
      additionalProperties: false
      required: [schemaVersion, producerVersion, producerInstanceId, sessionId, shardId, firstSequence, lastSequence, events]
      properties:
        schemaVersion: { type: integer, enum: [1] }
        producerVersion: { type: string }
        producerInstanceId: { type: string }
        sessionId: { type: string }
        shardId: { type: string }
        firstSequence: { type: integer, format: int64, minimum: 1 }
        lastSequence: { type: integer, format: int64, minimum: 1 }
        events: { type: array, maxItems: 256, items: { $ref: "#/components/schemas/DSTRuntimeEvent" } }
    DSTRuntimeEventBatchEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data: { $ref: "#/components/schemas/DSTRuntimeEventBatch" }
        error: { type: "null" }
        meta: { $ref: "#/components/schemas/Meta" }
    DSTRuntimeDiagnosticProfile:
      type: string
      enum: [summary, prefab, performance]
    DSTRuntimeDiagnosticInput:
      type: object
      additionalProperties: false
      required: [profile]
      properties:
        profile: { $ref: "#/components/schemas/DSTRuntimeDiagnosticProfile" }
        prefab: { type: string, pattern: "^[a-z0-9_]{1,80}$" }
        durationSeconds: { type: integer, minimum: 1, maximum: 5 }
        sampleLimit: { type: integer, minimum: 0, maximum: 50 }
    DSTRuntimeDiagnostic:
      type: object
      additionalProperties: false
      required: [schemaVersion, producerVersion, producerInstanceId, sessionId, shardId, sequence, requestId, profile, ok, code, message, result, completedAtUnix]
      properties:
        schemaVersion: { type: integer, enum: [1] }
        producerVersion: { type: string }
        producerInstanceId: { type: string }
        sessionId: { type: string }
        shardId: { type: string }
        sequence: { type: integer, format: int64, minimum: 1 }
        requestId: { type: string }
        profile: { $ref: "#/components/schemas/DSTRuntimeDiagnosticProfile" }
        ok: { type: boolean }
        code: { type: string }
        message: { type: string }
        result: { type: object, additionalProperties: true }
        completedAtUnix: { type: integer, format: int64, minimum: 1 }
    DSTRuntimeDiagnosticEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data: { $ref: "#/components/schemas/DSTRuntimeDiagnostic" }
        error: { type: "null" }
        meta: { $ref: "#/components/schemas/Meta" }
    DSTRuntimeReport:
      type: object
      additionalProperties: false
      required: [roomId, worldId, worldName, state, processRunning, healthState]
      properties:
        roomId: { type: string }
        worldId: { type: string }
        worldName: { type: string }
        state: { $ref: "#/components/schemas/DSTRuntimeInstallState" }
        version: { type: string }
        protocolVersion: { type: integer, minimum: 1 }
        changed: { type: boolean }
        backupPath: { type: string }
        message: { type: string }
        processRunning: { type: boolean }
        healthState: { $ref: "#/components/schemas/DSTRuntimeHealthState" }
        health: { $ref: "#/components/schemas/DSTRuntimeHealth" }
        healthMessage: { type: string }
    DSTRuntimeStatusEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data: { $ref: "#/components/schemas/DSTRuntimeStatus" }
        error: { type: "null" }
        meta: { $ref: "#/components/schemas/Meta" }
    DSTRuntimeStatusList:
      type: object
      additionalProperties: false
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/DSTRuntimeStatus" } }
        total: { type: integer, minimum: 0 }
    DSTRuntimeStatusListEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data: { $ref: "#/components/schemas/DSTRuntimeStatusList" }
        error: { type: "null" }
        meta: { $ref: "#/components/schemas/Meta" }
    DSTRuntimeReportList:
      type: object
      additionalProperties: false
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/DSTRuntimeReport" } }
        total: { type: integer, minimum: 0 }
    DSTRuntimeReportListEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data: { $ref: "#/components/schemas/DSTRuntimeReportList" }
        error: { type: "null" }
        meta: { $ref: "#/components/schemas/Meta" }
    DSTRuntimeBackup:
      type: object
      additionalProperties: false
      required: [id, worldId, createdAt, reason]
      properties:
        id: { type: string, pattern: '^state-[0-9]{8}T[0-9]{6}\.[0-9]{9}Z-[0-9a-f]{32}$' }
        worldId: { type: string }
        createdAt: { type: string, format: date-time }
        reason: { type: string, enum: [before-install, before-uninstall, before-rollback] }
    DSTRuntimeBackupList:
      type: object
      additionalProperties: false
      required: [items, total]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/DSTRuntimeBackup" } }
        total: { type: integer, minimum: 0 }
    DSTRuntimeBackupListEnvelope:
      type: object
      required: [data, error, meta]
      properties:
        data: { $ref: "#/components/schemas/DSTRuntimeBackupList" }
        error: { type: "null" }
        meta: { $ref: "#/components/schemas/Meta" }
    DSTRuntimeConfirmationInput:
      type: object
      additionalProperties: false
      required: [confirmation]
      properties:
        confirmation: { type: string, description: Must exactly equal the room name }
    DSTRuntimeRollbackInput:
      type: object
      additionalProperties: false
      required: [confirmation, backupId]
      properties:
        confirmation: { type: string, description: Must exactly equal the room name }
        backupId: { type: string, pattern: '^state-[0-9]{8}T[0-9]{6}\.[0-9]{9}Z-[0-9a-f]{32}$' }
...
security:
  - sessionCookie: []